Fix graph dates and rearrange code
